requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.0
python-dotenv==1.0.0
sentence-transformers==2.7.0
huggingface-hub==0.20.3
faiss-cpu==1.7.4
numpy==1.24.3
slowapi==0.1.9
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
transformers==4.36.0